% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allbyall}
\alias{allbyall}
\title{rowxrow or colxcol execution}
\usage{
allbyall(mat, margin = 2, fun)
}
\arguments{
\item{mat}{Matrix}

\item{margin}{Margin: col(2) or row(1)}

\item{fun}{Input function}
}
\description{
rowxrow or colxcol execution
}
