% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExpectedHets.R
\name{calcExpectedHets}
\alias{calcExpectedHets}
\title{Calculates expected number of heterozygous SNPs}
\usage{
calcExpectedHets(dbsnp_file, coverage = c(1), max_cnt = 2)
}
\arguments{
\item{dbsnp_file}{Path to dbSNP bed file (from setup_dbSNP.sh)}

\item{coverage}{Target coverage(s) and lambda param of rpois()}

\item{max_cnt}{Count to set all dbSNP at for max point (default=2)}
}
\value{
A matrix for the input coverages and the mean, mean^2 (for 2 samples 
overlapping), standard error heterozygous SNPs (Expected values) and
the number of SNPs
}
\description{
Takes a dbSNP file that is converted into the .bed format
by the setup_dbSNP.sh script and runs simulations to calculate the 
expected number of heterozygous SNPs in a single sample given a target
coverage fitted to a Poisson distribution.
}
