% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{combZygPos}
\alias{combZygPos}
\title{Zygosity combiner helper}
\usage{
combZygPos(zyg, zygpos, ret = "reduced")
}
\arguments{
\item{zyg}{data.frame object returned from fitHMM()$df}

\item{zygpos}{GRanges object that corresponds with zyg}

\item{ret}{return 'reduced' or a 'raw' state (bins or segments)}
}
\value{
A GRanges object that maps zygpos to zyg
}
\description{
Combines the zygosity data with the zygosity
position
}
