% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{fitHMM}
\alias{fitHMM}
\title{Fit an HMM}
\usage{
fitHMM(
  sample,
  het_cnt,
  states = 2,
  family = poisson(),
  nstart = 10,
  tiles = NULL,
  is.sim = FALSE,
  ret.raw = FALSE,
  multi = FALSE
)
}
\arguments{
\item{sample}{String containing sample ID in columns}

\item{het_cnt}{Dataframe for sample observations}

\item{states}{Integer for expected number of states}

\item{family}{Family of distribution to use (default=poisson())}

\item{nstart}{Number of starts to try out for EM (Default=10)}

\item{tiles}{GenomicRanges object for corresponding chrs and bins.  If none is given, it
will fill in the genomic poisition randomly (Default=NULL)}

\item{is.sim}{Boolean flag for whether data is simulation or not (Default=FALSE)}

\item{ret.raw}{Boolean flag to return the fitted model depmixS4 object}
}
\value{
Dataframe containing the emission proabbility for each state and a fitted HMM
}
\description{
Runs depmixs4 to fit an HMM to a set of observations
given an expected number of states. It will then configure the 
emission dataframe to be compatible with downstream analysis
}
