% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSnpZygosity.R
\name{genSnpZygosity}
\alias{genSnpZygosity}
\title{Simplifies the SNP Zygosity from MuTect}
\usage{
genSnpZygosity(sampleid, chrs = paste0("chr", c(1:22, "X", "Y")))
}
\arguments{
\item{sampleid}{Vector of sample IDs for the filename [chr].[sampleid].vcf}

\item{chrs}{Chrs for the filename [chr].[sampleid].vcf}
}
\value{
Matrix of parsed output values
}
\description{
Takes a single sample Mutect VCF that has force called all dbSNP
sites and reduces it to a single column of 1,2,3 values corresponding to 
REF.HOMOZYGOUS, HETEROZYGOUS, and ALT.HOMOZYGOUS respectively using Perl. 
If multiple samples are given, it will concatenate the samples into a single matrix.
}
