% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExpectedHets.R
\name{getExpectedN}
\alias{getExpectedN}
\title{Calculates the expected number of SNPs}
\usage{
getExpectedN(mu, se, n, genome_size = 3.2 * 10^9)
}
\arguments{
\item{mu}{Mean}

\item{se}{Standard error}

\item{n}{N}

\item{genome_size}{Default (3.2*10^9)}
}
\value{
95% confidence for number of SNPs given mu, se, and number of SNPs
95% confidence for genomic distance between those SNPs
}
\description{
Calculates the 95% confidence interval for the number of SNPs
given a total N, as well as the distance between these SNPs
}
