% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{metrics}
\alias{metrics}
\title{Multi-class performance metrics}
\usage{
metrics(predictions, reference)
}
\arguments{
\item{predictions}{vector of predictions}

\item{reference}{vector of actual classes}
}
\value{
2 element list of a dataframe containing micro and macro F1 scores with 
precision and accuracy
}
\description{
Get multi-class performance metrics in the form of
Precision, Accuracy, and micro/macro F1 metrics. The base metrics
are calculated from the caret packages confusionMatrix() function
}
\examples{
metrics(predictions=sample(1:4, 1000, replace=TRUE), reference=sample(1:4, 1000, replace=TRUE))
}
