% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseOutput}
\alias{parseOutput}
\title{Returns the tag from an execution}
\usage{
parseOutput(results, tag)
}
\arguments{
\item{results}{vector of printed statments}

\item{tag}{ID tag for the parsed output (e.g. Runtime)}
}
\value{
Data value for the parsed output
}
\description{
Looks for and parses the
'Runtime: 25' line in shell executions
}
