% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{parseSeg}
\alias{parseSeg}
\title{Parse CN seg files}
\usage{
parseSeg(
  segf,
  segcol = "event",
  start = "start",
  end = "end",
  state_frac = 0.9,
  min_frac = 0.05,
  get_tt = FALSE,
  get_seg = FALSE
)
}
\arguments{
\item{segf}{path to CN seg file}

\item{segcol}{character: column name for discrete CN state}

\item{start}{character: column name for interval start}

\item{end}{character: column name for interval end}

\item{state_frac}{numeric: minimal fraction of the genome occupied by CN states (default=0.90)}

\item{min_frac}{numeric: minimum segment size in fractions (default=0.05)}

\item{get_tt}{boolean: return a transition state probability matrix (in.dev)}

\item{get_seg}{boolean: return the sef file parsed}
}
\value{
A one or two element list:
  'state': integer number of states
  'tt': transition matrix
}
\description{
Parses a seg file with given chromosome-separated
intervals and a discrete CN state. It will identify the number of
states occupying `state_frac` of the genome, where each interval
is of at least `min_frac` in size.
}
