% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_viz.R
\name{plotCnZyg}
\alias{plotCnZyg}
\title{Cn-Zyg plotter}
\usage{
plotCnZyg(cn, zyg, tiles, cncol = "copy.number", ...)
}
\arguments{
\item{cn}{GRanges object with a discrete integer cncol column}

\item{zyg}{GRanges object with a discrete 'state' column of c('LOH', 'H1', etc...)}

\item{tiles}{Reference GRanges object from tileGenome()}

\item{cncol}{character: copy number column for cn GRanges object}

\item{...}{pass in for plot()}
}
\value{
data frame containing the segment-level CN and zygosity state
}
\description{
Plots the zygosity and cn state
}
