% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_viz.R
\name{plotHMM}
\alias{plotHMM}
\title{Plot zygosity HMM}
\usage{
plotHMM(model, est_cols, act_cols = NULL, xval = "bin")
}
\arguments{
\item{model}{Data frame output fitHmm() function with obs, states, and PP}

\item{est_cols}{2 element character vector for the estimated [state, state.label] columns}

\item{act_cols}{If known, 2 element character vector for the actual [state, state.label] columns}

\item{xval}{String for column ID corresponding to x-axis value (e.g. bin)}
}
\value{
a grid object for visualization
}
\description{
a 3-4 track visualization of an HMM fitted to a series of observed
heterozygous counts. The top track will be a simple count of heterozygous counts
along the genome. The middle track(s) will be the estimated state based on a
fitted HMM model. If the actual state is known, it will also include the visualization
for the actual state on top of the estimated states. The bottom track contains
line graphs for the posterior probability of each state.
}
