% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleSimilarity.R
\name{plotSampleSimilarity}
\alias{plotSampleSimilarity}
\title{Sim-N plotter}
\usage{
plotSampleSimilarity(sim_mat, n_mat, midpoint = 0.5)
}
\arguments{
\item{sim_mat}{Similarity matrix}

\item{n_mat}{Matrix of N's, same size and format as sim_mat}

\item{midpoint}{Midpoint value of colour range (Default=0.5)}
}
\description{
Plot the similarity and n-matrix
}
