% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturationCurve.R
\name{saturationCurve}
\alias{saturationCurve}
\title{Saturation curve estimations using asymptotic regression models}
\usage{
saturationCurve(dat, pred = NULL, S = NULL, Xin = NULL)
}
\arguments{
\item{dat}{Two column data frame where dat[,1] is X values and dat[,2] is response}

\item{pred}{Series of X values to predict on (Default=NULL)}

\item{S}{Saturation point to calculate the X value for (Default=NULL)}

\item{Xin}{X value to calculate the saturation point for (Default=NULL)}
}
\value{
A list of the model data, coefficients, and returned values if pred, Xin, or
S are given
}
\description{
Calculates the nonlinear model fit using an 
Asymptotic Regression Model from and non-linear least squares
from the nlstools::nls and stats::SSasymp functions, where the
following parameters are being estimated:
 A = Response at 100% saturation (horizontal asymptote)
 R = Response when x = 0
 l = Natural logarithm of the rate constant
 
 Given an input of X values to predict on (pred), this function will 
 return the predicted Response based on the model. Or the 
 corresponding X for a given saturation (S), or a Saturation 
 associated with a given X.
}
