% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(
  N,
  cnts = c(loh = 10, het = 15),
  switch.prob = 0.1,
  family = "poisson"
)
}
\arguments{
\item{N}{Integer value for number of elements to randomly generate}

\item{cnts}{Named integer vector for mean/lambda value for that state}

\item{switch.prob}{Numeric value for transition probability (Default=0.1)}

\item{family}{Family of distributions (Default=poission)}
}
\value{
A Dataframe consisting of states and the simulated observations from each state
}
\description{
Simulates HMM data using a given family of distributions, 
an input number of states, and a transition probability.
}
\examples{
set.seed(1234)
simulate(100, cnts=setNames(c(10, 20), c('loh', 'het')), 
         switch.prob=0.05, family='poisson')
}
