% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{toBed}
\alias{toBed}
\title{GRanges to Bed}
\usage{
toBed(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\value{
Dataframe
}
\description{
simple function to convert a GRanges object to a data
frame in .bed file format
}
