% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleSimilarity.R
\name{getSampleSimilarity}
\alias{getSampleSimilarity}
\title{Generates a similarity matrix}
\usage{
getSampleSimilarity(
  sample_matrix,
  samples = NULL,
  matchmode = "autosome",
  similarityFun = NULL,
  rm_nocov = FALSE,
  rem_ref = TRUE
)
}
\arguments{
\item{sample_matrix}{character: path to csv or vcf file to run similarity match on}

\item{samples}{character: Comma-separated list of ordered samples for the matrix}

\item{matchmode}{character: Either 'autosome' or 'chrM' (Default=autosome)}

\item{similarityFun}{Similarity function with two parameters, 'i' and 'j'.}

\item{rm_nocov}{If no filtering process took place prior, setting this to TRUE
will remove all SNPs where no sample have any coverage (Default=FALSE)}

\item{rem_ref}{boolean: Removes reference matches in chrM sample matching (0/0 and 
0/0) (Default=TRUE)}
}
\value{
Returns a list of similarity matrices, matrix of number of SNPs found between
two samples, and a matrix of number of heterozygous SNPs between two samples
}
\description{
Takes a path to a directory that contains
the [CHR]_filt.snps files and processes each to generate
a similarity matrix using a given similarityFun.  If no
similarityFun is given, it generates one using the default
Jaccard metric.
}
