% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleSimilarity.R
\name{getSampleSimilarity}
\alias{getSampleSimilarity}
\title{Title}
\usage{
getSampleSimilarity(filt_dir, samples, similarityFun = NULL)
}
\arguments{
\item{filt_dir}{to fill in}

\item{samples}{to fill in}

\item{similarityFun}{to fill in}
}
\value{
to fill in
}
\description{
Title
}
