% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{sanitizeData}
\alias{sanitizeData}
\title{Sanitize the data}
\usage{
sanitizeData(x, winsorize_p = 0.01)
}
\arguments{
\item{x}{Numeric vector}

\item{winsorize_p}{Numeric value for probabilty to winsorize at (Default = 0.01)}
}
\value{
Numeric vector of median normalized x in log space, followed by a 0.01,0.099 winsorization
}
\description{
If data is distributed according to a poisson distribution, median normalize
the data and log transform to make it more gaussian like
}
\examples{
sanitizeData(rpois(100, 10))
}
